// src/models/enums/PluginActions.ts
/* eslint-disable no-unused-vars */
// eslint-disable-next-line no-shadow
enum PluginActions {
  onTouchSlider = 'onTouchSlider',
  onTouchHandler = 'onTouchHandler',
  onMoveHandler = 'onMoveHandler',
  onStopMoving = 'onStopMoving',
  onChangeState = 'onChangeState',
  onChangeValue = 'onCangeValue',
  onDestroy = 'onDestroy',
}

export default PluginActions;